refactored database connection and following MVC model
