+++ templates/signup.html
@@ -16,7 +16,7 @@
           </div>
         <div class="form-group">
         <label for="email">Email</label>
-        <input type="text" class="form-control" id="email" name="email" placeholder="email">
+        <input type="email" class="form-control" id="email" name="email" placeholder="email">
       </div>
       <div class="form-group">
         <label for="password">Password</label>
